+++ src/app/api/admin/deleteschedule/route.ts
@@ -2,18 +2,26 @@ import { NextRequest, NextResponse } from "next/server";
 
 export async function DELETE(req: NextRequest) {
 	try {
-		const { _id } = await req.json();
+		const { searchParams } = new URL(req.url);
+		const scheduleId = searchParams.get("scheduleId");
+		
+		if (!scheduleId) {
+			return NextResponse.json(
+				{ message: "Schedule ID is required" },
+				{ status: 400 }
+			);
+		}
+
 		const token = req.headers.get("authorization") || "";
 
 		const response = await fetch(
-			`${process.env.NEXT_PUBLIC_BACKAPI_URL}/api/admin/delete/schedule`,
+			`${process.env.NEXT_PUBLIC_BACKAPI_URL}/api/admin/schedules/${scheduleId}`,
 			{
 				method: "DELETE",
 				headers: {
 					"Content-Type": "application/json",
 					Authorization: token,
 				},
-				body: JSON.stringify({ _id }),
 			}
 		);
